[fix] type longitude et latitude

+++ src/main/java/com/mycompany/roadtripplanner/dtos/position/PositionGetDTO.java
@@ -16,7 +16,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @AllArgsConstructor
 public class PositionGetDTO {
     private String id;
-    private Long longitude;
-    private Long latitude;
+    private double longitude;
+    private double latitude;
 
 }
